refactor(blorktools): css3d prerender
